requirements.txt
azure-common==1.1.3
azure-nspkg==1.0.0
# itsdangerous==2.0
# Jinja2==3.0
# MarkupSafe==2.0
python-dateutil==2.7.3
requests==2.20.0
six==1.12.0
Werkzeug==2.0
azure-mgmt-core==1.3.0
azure-mgmt-storage==18.0.0
azure-storage-blob==12.2.0
azure-identity==1.6.0
azure-keyvault-secrets==4.3.0
flask==2.0.2
pandas==1.3.3